hyundai_kona_ev/bench_kona/requirements.txt
python-can
pyside6>=6.7.0<6.8
